Refactoring, add button
