+++ app/src/main/java/com/uzlov/moviefind/HostActivity.kt
@@ -4,12 +4,18 @@ import android.os.Bundle
 import android.view.Menu
 import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
+import com.uzlov.moviefind.fragments.HomeFragment
 
 class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
+
+        supportFragmentManager
+            .beginTransaction()
+            .replace(R.id.fragment_container, HomeFragment.newInstance())
+            .commit()
     }
 
     override fun onCreateOptionsMenu(menu: Menu): Boolean {
