+++ app/src/main/java/com/uzlov/moviefind/fragments/HomeFragment.kt
@@ -1,0 +1,32 @@
+package com.uzlov.moviefind.fragments
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.Button
+import android.widget.Toast
+import com.uzlov.moviefind.R
+
+class HomeFragment : Fragment() {
+
+    companion object{
+        fun newInstance() : HomeFragment {
+            return HomeFragment()
+        }
+    }
+    override fun onCreateView(
+            inflater: LayoutInflater, container: ViewGroup?,
+            savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_home, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        view.findViewById<Button>(R.id.button_first).setOnClickListener {
+            Toast.makeText(requireContext(), getString(R.string.thankyou), Toast.LENGTH_SHORT).show()
+        }
+    }
+}
