+++ app/src/main/res/values-night/strings.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="okay">Круто!</string>
+    <string name="thankyou">Спасибо</string>
+</resources>
