+++ app/src/main/res/values/strings.xml
@@ -1,12 +1,9 @@
 <resources>
     <string name="app_name">MovieFind</string>
-    <string name="action_settings">Settings</string>
-    <!-- Strings used for fragments for navigation -->
-    <string name="first_fragment_label">First Fragment</string>
-    <string name="second_fragment_label">Second Fragment</string>
-    <string name="next">Next</string>
-    <string name="previous">Previous</string>
+    <string name="action_settings">Настройки</string>
 
-    <string name="hello_first_fragment">Hello first fragment</string>
-    <string name="hello_second_fragment">Hello second fragment. Arg: %1$s</string>
+
+    <string name="hello_first_fragment">Скоро тут будут фильмы…</string>
+    <string name="okay">Круто!</string>
+    <string name="thankyou">Спасибо</string>
 </resources>
